+++ RememberRoot.h
@@ -14,7 +14,7 @@ namespace Remember
    public:
       inline static void setUnsynced();
       inline static bool isUnsynced();
-      inline size_t compileDataSize() const;
+      inline uint64_t compileDataSize() const;
       inline DataVector get() const;
       inline void set(const DataVector& data);
