+++ RememberValueList.h
@@ -21,12 +21,12 @@ namespace Remember
       inline DataType& operator[](const uint16_t index);
       inline const DataType& operator[](const uint16_t index) const;
       inline void append(const DataType& value);
-      inline size_t size() const;
+      inline uint64_t size() const;
       inline void clear();
 
    protected:
       inline void write(DataVector& data) const override;
-      inline void read(const DataVector& data, size_t& cursor) override;
+      inline void read(const DataVector& data, uint64_t& cursor) override;
 
    protected:
       std::vector<DataType> members;
